Deduplicate data points based on cell offset
Deduplicates data points based on the cell offset. If the type of the
offset differs, usually where a normal int would overflow and a long is
used, we may end up writing duplicate metrics (with the format flags
inticating the type).

This change strips out the format flags and parses the integer correctly
every time now, regardless of the size of the variable.

http://opentsdb.net/docs/build/html/user_guide/backends/hbase.html#data-point-columns)

Co-authored-by: Neil Fordyce <neil.fordyce@skyscanner.net>